requirements.txt
flask
astropy
flask-restplus
pyyaml
gunicorn
bokeh
matplotlib
numpy
requests